requirements.txt
pytorch
gdal
rasterio
scikit-learn
numpy
tensorboard
torchvision
scikit-image
pandas
pytables
tqdm